chore: move project name config to common
to get rid of a dependency cycle
